COMP: Fix various warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=17140
git-svn-id: http://svn.slicer.org/Slicer4/trunk@17140 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRMLLogic/vtkMRMLSliceLayerLogic.cxx
@@ -497,10 +497,10 @@ void vtkMRMLSliceLayerLogic::UpdateImageDisplay()
     return;
     }
   
-  int oldSliceMTime = this->Slice->GetMTime();
-  int oldReSliceMTime = this->Reslice->GetMTime();
-  int oldAssign = this->AssignAttributeTensorsFromScalars->GetMTime();
-  int oldLabel = this->LabelOutline->GetMTime();
+  unsigned long oldSliceMTime = this->Slice->GetMTime();
+  unsigned long oldReSliceMTime = this->Reslice->GetMTime();
+  unsigned long oldAssign = this->AssignAttributeTensorsFromScalars->GetMTime();
+  unsigned long oldLabel = this->LabelOutline->GetMTime();
   
   if ( (this->VolumeNode->GetImageData() && labelMapVolumeDisplayNode) ||
        (scalarVolumeDisplayNode && scalarVolumeDisplayNode->GetInterpolate() == 0))
